requirements.txt
langchain==0.0.274
gpt4all
chromadb==0.4.7
urllib3==2.0.4
python-dotenv==1.0.0
unstructured==0.10.8
extract-msg==0.45.0
tabulate==0.9.0
pandoc==2.3
pypandoc==1.11
tqdm==4.66.1
sentence_transformers==2.2.2
pypdf
streamlit
streamlit-extras
python-decouple
streamlit_option_menu